+++ .github/workflows/python.yml
@@ -12,7 +12,7 @@ jobs:
         os: [ubuntu-latest, macos-13, windows-latest]
         # Lowest and highest, no version specified so that 
         # new releases get automatically tested against
-        version: [{torch: torch==1.10, python: "3.8"}, {torch: torch, python: "3.12"}]
+        version: [{torch: torch==1.10, python: "3.8", pyfeature: "38"}, {torch: torch, python: "3.12", pyfeature: "311"}]
           # TODO this would include macos ARM target.
           # however jax has an illegal instruction issue
           # that exists only in CI (probably difference in instruction support).
@@ -59,31 +59,36 @@ jobs:
 
       - name: Install
         run: |
-          pip install -U pip
-          pip install .[numpy,tensorflow]
+          pip install -U pip maturin
+          # pip install .[numpy,tensorflow]
+          maturin develop --features ${{ matrix.version.pyfeature }},pyo3/extension-module --extras numpy,tensorflow
           pip install ${{ matrix.version.torch }}
 
       - name: Install (jax, flax)
         if: matrix.os != 'windows-latest'
         run: |
-          pip install .[jax]
+          # pip install .[jax]
+          maturin develop --features ${{ matrix.version.pyfeature }},pyo3/extension-module --extras jax
         shell: bash
 
       - name: Install (mlx)
         if: matrix.os == 'macos-latest'
         run: |
-          pip install .[mlx]
+          # pip install .[mlx]
+          maturin develop --features ${{ matrix.version.pyfeature }},pyo3/extension-module --extras mlx
         shell: bash
 
       - name: Check style
         run: |
-          pip install .[quality]
+          # pip install .[quality]
+          maturin develop --features ${{ matrix.version.pyfeature }},pyo3/extension-module --extras quality
           black --check --line-length 119 --target-version py35 py_src/safetensors tests
 
       - name: Run tests
         run: |
           cargo test
-          pip install .[testing]
+          # pip install .[testing]
+          maturin develop --features ${{ matrix.version.pyfeature }},pyo3/extension-module --extras testing
           pytest -sv tests/
 
   test_s390x_big_endian:
